requirements.txt
streamlit==1.24.0
pandas==2.0.0
pandasai==0.0.9
langchain-groq==0.1.0
seaborn==0.12.2
matplotlib==3.7.1
plotly==5.10.0
altair==4.2.0
python-dotenv==1.0.0
